Added .help (cmd) functionality to .help command
Added a JSON file, required fs module. Added a variable called cmd which is equal to the 0th argument (1st). A check is made if the command supplied is in the accepted list of commands to get help for, if it fails, a DM is sent to the user containing all commands that can be used with a randomized example. Embed is used to send out the help about the specific command.